Rework Fusion: link units as parents

+++ schema/src/fusion.rs
@@ -2,13 +2,13 @@ use super::*;
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Default, Hash)]
 pub struct UnitActionRef {
-    pub unit: u8,
+    pub unit: u64,
     pub trigger: u8,
     pub action: u8,
 }
 
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Serialize, Deserialize, Default)]
 pub struct UnitTriggerRef {
-    pub unit: u8,
+    pub unit: u64,
     pub trigger: u8,
 }
